+++ lib/module/index.dart
@@ -8,15 +8,14 @@ export 'login/login.dart';
 export 'my-operation/my-operation-service.dart';
 export 'my-operation/my-operation.dart';
 export 'my-operation/operations.dart';
+export 'my-operation/reservation-create.dart';
+export 'my-operation/reservation.dart';
 export 'my-programing/my-programing-service.dart';
 export 'my-programing/my-programing.dart';
 export 'my-programing/sport-detail/sport-details.dart';
 export 'payment/payment-model.dart';
 export 'payment/payment-service.dart';
-export 'payment/payment.dart';
 export 'profile/profile-module/measurements.dart';
 export 'profile/profile-module/member-type.dart';
-export 'profile/profile-module/reservation-create.dart';
-export 'profile/profile-module/reservation.dart';
 export 'profile/profile-service.dart';
 export 'profile/profile.dart';
