+++ lib/module/my-operation/reservation-create.dart
@@ -3,7 +3,7 @@ import 'package:elektra_fit/widget/index.dart';
 import 'package:flutter/material.dart';
 import 'package:lottie/lottie.dart';
 
-import '../../../global/global-models.dart';
+import '../../global/global-models.dart';
 
 class ReservationCreate extends StatefulWidget {
   ReservationCreate({required this.selectedHours, required this.resStart, required this.spaService, super.key});
@@ -17,11 +17,11 @@ class ReservationCreate extends StatefulWidget {
 }
 
 class _ReservationCreateState extends State<ReservationCreate> {
-  final service = GetIt.I<ProfileService>();
+  final service = GetIt.I<MyOperationsService>();
 
-  TextEditingController _nameAndSurname = TextEditingController();
-  TextEditingController _email = TextEditingController();
-  TextEditingController _phone = TextEditingController();
+  final TextEditingController _nameAndSurname = TextEditingController();
+  final TextEditingController _email = TextEditingController();
+  final TextEditingController _phone = TextEditingController();
   final _formKey = GlobalKey<FormState>();
 
   Future<void> reservationCreate() async {
@@ -29,7 +29,8 @@ class _ReservationCreateState extends State<ReservationCreate> {
     if (validationResult) {
       service.reservationCreate(_nameAndSurname.text, _phone.text, widget.resStart, widget.spaService, service.paymentType$.value).then((value) {
         if (value!.result) {
-          reservationSuccesfullWidget(_email.text, _phone.text, _nameAndSurname.text, widget.spaService, widget.resStart, widget.selectedHours, service.resId$.value);
+          reservationSuccesfullWidget(
+              _email.text, _phone.text, _nameAndSurname.text, widget.spaService, widget.resStart, widget.selectedHours, service.resId$.value);
         } else {
           kShowBanner(BannerType.ERROR, value.message, context);
         }
@@ -76,14 +77,20 @@ class _ReservationCreateState extends State<ReservationCreate> {
                     Column(mainAxisAlignment: MainAxisAlignment.start, crossAxisAlignment: CrossAxisAlignment.start, children: [
                       Text("Service Information".tr(), style: kMontserrat17),
                       const Divider(color: Colors.black45),
-                      Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [Text("Service Name".tr(), style: kProxima17), Text(widget.spaService.product, style: kProxima17)]),
                       Row(
                           mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [Text("Service Price".tr(), style: kProxima17), Text("${widget.spaService.price} ${widget.spaService.currency}", style: kProxima17)]),
+                          children: [Text("Service Name".tr(), style: kProxima17), Text(widget.spaService.product, style: kProxima17)]),
+                      Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
+                        Text("Service Price".tr(), style: kProxima17),
+                        Text("${widget.spaService.price} ${widget.spaService.currency}", style: kProxima17)
+                      ]),
+                      Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
+                        Text("Reservation Date".tr(), style: kProxima17),
+                        Text(DateFormat("dd MMM yyyy").format(widget.resStart), style: kProxima17)
+                      ]),
                       Row(
                           mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [Text("Reservation Date".tr(), style: kProxima17), Text(DateFormat("dd MMM yyyy").format(widget.resStart), style: kProxima17)]),
-                      Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [Text("Reservation Time".tr(), style: kProxima17), Text(widget.selectedHours.toString(), style: kProxima17)]),
+                          children: [Text("Reservation Time".tr(), style: kProxima17), Text(widget.selectedHours.toString(), style: kProxima17)]),
                       SizedBox(height: W / 30),
                       Text("Person Information".tr(), style: kMontserrat17),
                       const Divider(color: Colors.black45),
@@ -103,7 +110,8 @@ class _ReservationCreateState extends State<ReservationCreate> {
                         },
                       ),
                       SizedBox(height: W / 40),
-                      CTextFormField(_email, "Email".tr(), keyboardType: TextInputType.emailAddress, textInputAction: TextInputAction.go, onchange: (v) {
+                      CTextFormField(_email, "Email".tr(), keyboardType: TextInputType.emailAddress, textInputAction: TextInputAction.go,
+                          onchange: (v) {
                         member$.value?.first.profile.email = v;
                       }, validator: (value) {
                         if (value.isEmpty) {
@@ -112,7 +120,8 @@ class _ReservationCreateState extends State<ReservationCreate> {
                         return null;
                       }),
                       SizedBox(height: W / 40),
-                      CTextFormField(_phone, "Phone Number".tr(), keyboardType: TextInputType.phone, textInputAction: TextInputAction.done, validator: (value) {
+                      CTextFormField(_phone, "Phone Number".tr(), keyboardType: TextInputType.phone, textInputAction: TextInputAction.done,
+                          validator: (value) {
                         if (value.isEmpty) {
                           return 'Please enter your phone number.'.tr();
                         }
@@ -135,7 +144,8 @@ class _ReservationCreateState extends State<ReservationCreate> {
     );
   }
 
-  Future<dynamic> reservationSuccesfullWidget(String email, String phone, String fullName, SpaService spaService, DateTime resStart, String selectedHours, int resId) {
+  Future<dynamic> reservationSuccesfullWidget(
+      String email, String phone, String fullName, SpaService spaService, DateTime resStart, String selectedHours, int resId) {
     return showModalBottomSheet(
       context: context,
       isScrollControlled: true,
@@ -146,14 +156,15 @@ class _ReservationCreateState extends State<ReservationCreate> {
             stream: isLoading$.stream,
             builder: (context, snapshot) {
               return Container(
-                alignment: Alignment.bottomRight,
-                height: H * 0.6,
-                padding: paddingAll10,
-                child: isLoading$.value
-                    ? Center(child: CircularProgressIndicator(color: config.primaryColor))
-                    : Column(
-                        children: [
-                          Container(alignment: Alignment.center, child: Lottie.asset('assets/animations/success.json', width: W / 3, height: W / 3, fit: BoxFit.cover)),
+                  alignment: Alignment.bottomRight,
+                  height: H * 0.6,
+                  padding: paddingAll10,
+                  child: isLoading$.value
+                      ? Center(child: CircularProgressIndicator(color: config.primaryColor))
+                      : Column(children: [
+                          Container(
+                              alignment: Alignment.center,
+                              child: Lottie.asset('assets/animations/success.json', width: W / 3, height: W / 3, fit: BoxFit.cover)),
                           SizedBox(width: W / 40),
                           Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
                             Text("Reservation Details".tr(), style: kMontserrat17),
@@ -163,13 +174,19 @@ class _ReservationCreateState extends State<ReservationCreate> {
                                 builder: (context, snapshot) {
                                   return Row(
                                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                    children: [Text("Reservation Number".tr(), style: kProxima17), Text(service.resId$.value.toString(), style: kProxima17)],
+                                    children: [
+                                      Text("Reservation Number".tr(), style: kProxima17),
+                                      Text(service.resId$.value.toString(), style: kProxima17)
+                                    ],
                                   );
                                 }),
                             SizedBox(width: W / 60),
                             Row(
                               mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                              children: [Text("Reservation Date".tr(), style: kProxima17), Text(DateFormat("dd MMM yyyy").format(resStart), style: kProxima17)],
+                              children: [
+                                Text("Reservation Date".tr(), style: kProxima17),
+                                Text(DateFormat("dd MMM yyyy").format(resStart), style: kProxima17)
+                              ],
                             ),
                             SizedBox(width: W / 60),
                             Row(
@@ -195,17 +212,17 @@ class _ReservationCreateState extends State<ReservationCreate> {
                             children: [Text("Email".tr(), style: kProxima17), Text(email, style: kProxima17)],
                           ),
                           SizedBox(width: W / 60),
-                          Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [Text("Phone Number".tr(), style: kProxima17), Text(phone, style: kProxima17)]),
+                          Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [Text("Phone Number".tr(), style: kProxima17), Text(phone, style: kProxima17)]),
                           const Divider(),
                           Padding(
-                            padding: EdgeInsets.only(bottom: MediaQuery.of(context).padding.bottom),
-                            child: Row(
-                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                              children: [
+                              padding: EdgeInsets.only(bottom: MediaQuery.of(context).padding.bottom),
+                              child: Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
                                 Expanded(
                                     child: InkWell(
                                   onTap: () {
-                                    Navigator.push(context, RouteAnimation.createRoute(Reservation(), 1, 0));
+                                    Navigator.push(context, RouteAnimation.createRoute(MyOperationDetail(), 1, 0));
                                   },
                                   child: Container(
                                       alignment: Alignment.center,
@@ -229,12 +246,8 @@ class _ReservationCreateState extends State<ReservationCreate> {
                                       decoration: BoxDecoration(border: Border.all(color: config.primaryColor), borderRadius: borderRadius10),
                                       child: Text("My Reservations".tr(), style: kProxima18.copyWith(color: config.primaryColor))),
                                 )),
-                              ],
-                            ),
-                          )
-                        ],
-                      ),
-              );
+                              ]))
+                        ]));
             });
       },
     );
