--- lib/module/payment/payment.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-class Payment extends StatefulWidget {
-  const Payment({super.key});
-
-  @override
-  State<Payment> createState() => _PaymentState();
-}
-
-class _PaymentState extends State<Payment> {
-  @override
-  Widget build(BuildContext context) {
-    return SingleChildScrollView();
-  }
-}
