+++ lib/widget/tabBar.dart
@@ -11,7 +11,7 @@ class CTabBar extends StatefulWidget {
 
 class _CTabBarState extends State<CTabBar> {
   int currentTab = 0;
-  final List<Widget> screens = [const Home(), const Qr(), MyOperations(), const MyPrograming(), const Profile()];
+  final List<Widget> screens = [const Home(), const Qr(), const MyOperations(), const MyPrograming(), const Profile()];
   final List<IconData> icons = [Icons.home, Icons.qr_code, Icons.add_business_outlined, Icons.list, Icons.person];
 
   Widget currentScreen = const Home();
@@ -30,69 +30,5 @@ class _CTabBarState extends State<CTabBar> {
               });
             }),
         body: screens[currentTab]);
-  }
-
-  //   return Scaffold(
-  //     body: screens[currentTab],
-  //     floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-  //     bottomNavigationBar: BottomAppBar(
-  //       height: W / 5.8,
-  //       notchMargin: 10,
-  //       padding: paddingAll10,
-  //       elevation: 0,
-  //       surfaceTintColor: Colors.yellowAccent,
-  //       color: Colors.black.withOpacity(0.2),
-  //       shape: CircularNotchedRectangle(),
-  //       child: Row(
-  //         mainAxisAlignment: MainAxisAlignment.spaceAround,
-  //         children: [
-  //           _buildTabItem(0, "assets/icon/tab-icon/home-selected.png", "assets/icon/tab-icon/home-unselected.png", "Home"),
-  //           _buildTabItem(1, "assets/icon/qr-icon.png", "assets/icon/qr-icon.png", "Qr"),
-  //           _buildTabItem(2, "assets/icon/tab-icon/my-programing-selected.png", "assets/icon/tab-icon/my-programing-unselected.png", "Activity"),
-  //           _buildTabItem(3, "assets/icon/tab-icon/profile-selected.png", "assets/icon/tab-icon/profile-unselected.png", "Profile"),
-  //         ],
-  //       ),
-  //     ),
-  //   );
-  // }
-  //
-  // Widget _buildTabItem(int index, String selectedImageAsset, String unselectedImageAsset, String text) {
-  //   return InkWell(
-  //     onTap: () {
-  //       _onTabTapped(index);
-  //     },
-  //     child: Container(
-  //       width: MediaQuery.of(context).size.width / 4.3,
-  //       child: Column(
-  //         crossAxisAlignment: CrossAxisAlignment.center,
-  //         mainAxisAlignment: MainAxisAlignment.center,
-  //         children: [
-  //           Container(
-  //             alignment: Alignment.center,
-  //             height: MediaQuery.of(context).size.width / 29,
-  //             width: MediaQuery.of(context).size.width / 20,
-  //             child: Image.asset(
-  //               currentTab == index ? selectedImageAsset : unselectedImageAsset,
-  //               color: currentTab == index ? config.IconPrimaryColor : Colors.black87,
-  //             ),
-  //           ),
-  //           Text(
-  //             text,
-  //             style: TextStyle(
-  //               color: currentTab == index ? config.IconPrimaryColor : Colors.black87,
-  //               fontFamily: "Proxima",
-  //               fontSize: 13,
-  //             ),
-  //           ),
-  //         ],
-  //       ),
-  //     ),
-  //   );
-  // }
-
-  void _onTabTapped(int index) {
-    setState(() {
-      currentTab = index;
-    });
   }
 }
